requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==4.0.0
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.1.2
django-quill-editor==0.1.16
django-use-email-as-username==1.1.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
Paperboy==1.0.1
psycopg2==2.8.5
-e git+https://github.com/gbozee/django-paystack.git@24df1c9ca209031e0ec940ff8552d45ca6b55421#egg=pypaystack
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
